Am adaugat explozii pentru bombtrap. Momentan exploziile in lant dau eroare. Voi rezolva asta in urmatorul commit. Suspectez ca eroare este cauzata de memorie manevrata necorespunzator. Voi adauga si destructori virtuali pentru toate clasele din ierarhie. Posibil sa incerc pe viitor si o arie de explozie circulara a bombelor. Va trebui sa rezolv si confuziile de coordonate din Map.

+++ DeLaGaze/Block.cpp
@@ -1,6 +1,6 @@
 #include "Block.h"
 
-Block::Block(std::pair<int, int> pos) :
+Block::Block(std::pair<size_t, size_t> pos) :
 	m_pos{pos}
 {
 }
