+++ DeLaGaze/Block.h
@@ -3,7 +3,7 @@
 
 class Block : public Object {
 public:
-	Block(std::pair<int, int> pos);
+	Block(std::pair<size_t, size_t> pos);
 	virtual void Render() override =0;
 	//OR
 	//void render(int x, int y) override;
@@ -11,7 +11,7 @@ public:
 	virtual void Print() const override =0;
 	virtual crow::json::wvalue toJson() override = 0;
 protected:
-	std::pair<int, int> m_pos;
+	std::pair<size_t, size_t> m_pos;
 private:
 
 };
