+++ DeLaGaze/Pathway.h
@@ -2,7 +2,7 @@
 #include "Block.h"
 class Pathway: public Block{
 public:
-	Pathway(std::pair<int, int> pos);
+	Pathway(std::pair<size_t, size_t> pos);
 	ObjectType GetType() const override { return ObjectType::Pathway; };
 	void Render() override;
 	void Print() const override;
