+++ DeLaGaze/Wall.h
@@ -2,7 +2,7 @@
 #include "Block.h"
 class Wall:public Block{
 public:
-	Wall(std::pair<int, int> pos);
+	Wall(std::pair<size_t, size_t> pos);
 	virtual void Render() override = 0;
 	virtual void Print() const override = 0;
 	virtual crow::json::wvalue toJson() override = 0;
